refactor: 합칠 수 있는 컨트롤러를 합침
