--- app/src/main/java/camp/nextstep/controller/LoginViewController.java
@@ -1,25 +1,0 @@
-package camp.nextstep.controller;
-
-import com.interface21.context.stereotype.Controller;
-import com.interface21.web.bind.annotation.RequestMapping;
-import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class LoginViewController {
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public ModelAndView showLogin(final HttpServletRequest req, final HttpServletResponse res) {
-        return UserSession.getUserFrom(req.getSession())
-                          .map(user -> {
-                              log.info("logged in {}", user.getAccount());
-                              return ModelAndView.jspView("redirect:/index.jsp");
-                          })
-                          .orElse(ModelAndView.jspView("/login.jsp"));
-    }
-}
