--- app/src/main/java/camp/nextstep/controller/RegisterViewController.java
@@ -1,20 +1,0 @@
-package camp.nextstep.controller;
-
-import com.interface21.context.stereotype.Controller;
-import com.interface21.web.bind.annotation.RequestMapping;
-import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.view.ViewUtil;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@Controller
-public class RegisterViewController {
-    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
-    public ModelAndView show(final HttpServletRequest req, final HttpServletResponse res) {
-        View view = ViewUtil.createViewByViewName("/register.jsp");
-
-        return new ModelAndView(view);
-    }
-}
